+++ core/printer/prettyprint/load.php
@@ -7,6 +7,21 @@ if ( ! defined( 'ABSPATH' ) ) {
 use Dev4Press\Plugin\DebugPress\Printer\PrettyPrint\PrettyPrint;
 
 /**
+ * Pretty print replacement that outputs the provided variables for debugging purposes.
+ *
+ * @param mixed ...$value One or more variables to be dumped for debugging.
+ *
+ * @return void
+ */
+function debugpress_p( ...$value ) {
+	foreach ( $value as $v ) {
+		$n = PrettyPrint::instance( $v );
+
+		$n->render();
+	}
+}
+
+/**
  * Main `print_r` pretty print replacement that can pretty print and format (almost) anything you want including
  * objects with support for reflection analysis.
  *